move loading CasC model from yaml under yaml util package made public for easier re-use in another context
